Refactor class methods to util files
